RN-268 Fixed teams list not always rerendering when team member changes
